refactor: error correction
