+++ src/app/providers/storeProvider/index.ts
@@ -1,3 +1,5 @@
+export { store } from './config/store';
+
 export { StoreProvider } from './ui/storeProvider';
 
 export type { AppDispatch, RootState } from './config/store';
