+++ src/entities/SelectedMovies/index.ts
@@ -4,6 +4,6 @@ export {
   clearStateMovies,
 } from './model/slices/selectedMoviesSlices';
 
-export { getSelectedMovies } from './model/selectors/selectors';
+export { getSelectedMovies } from './model/selectors/selectedMoviesSelectors';
 
 export type { SelectedMovieSchema } from './model/types/selectedMoviesTypes';
