+++ src/features/CreateMovieCardList/ui/MoviesCardList.css
@@ -5,6 +5,7 @@
   width: 75%;
   border-top: 1px solid grey;
   margin: 30px auto;
+  padding-bottom: 70px;
 }
 
 .characterCardList {
