+++ src/features/UseSelectedMovies/ui/SelectedMoviesModal.css
@@ -1,23 +1,24 @@
 .modal {
   position: fixed;
-  bottom: 5%;
-  right: 70px;
+  bottom: 0;
+  left: 50%;
   display: flex;
-  flex-direction: column;
   align-items: center;
+  justify-content: center;
   gap: 15px;
-  width: 350px;
-  height: 45px;
+  width: 60%;
+  height: 0px;
   font-size: 24px;
   border-radius: 5px;
   border: 1px solid white;
   background-color: rgb(63, 62, 62);
-  transition: height 0.5s ease-in-out;
+  transform: translate(-50%, 0);
+  transition: height 0.3s ease-in-out;
   z-index: 20;
 }
 
 .modal_open {
-  height: 170px;
+  height: 70px;
   padding: 10px;
 }
 
@@ -43,11 +44,11 @@
 
 @keyframes height {
   0% {
-    height: 45px;
+    height: 0px;
     padding: 0px;
   }
   100% {
-    height: 170px;
+    height: 70px;
     padding: 10px;
   }
 }
