+++ src/shared/test/index.ts
@@ -1,0 +1,1 @@
+export { testMovie, testMovies } from './mockData';
